requirements.txt
python-telegram-bot>=21.0,<22.0
Pillow>=10.4.0
pytesseract==0.3.10
requests>=2.31.0
python-dotenv>=1.0.0
Flask>=2.3.0
